Re-org
- add module `ibox`
  - `via` -> `ibox`
  - simplify `ibox` classes
- rename commands
